nargo_expand
Module items

This module is responsible for building a list of items that represent the final code that is going to be monomorphized and turned into SSA. This final code has all macros expanded and is mainly gathered from data inside NodeInterner, modules in DefMaps and function bodies from HirExpressions.

Structs

DataType ğŸ”’

Impl ğŸ”’

Import ğŸ”’

ItemBuilder ğŸ”’

Module ğŸ”’

Trait ğŸ”’

TraitImpl ğŸ”’

Enums

Item ğŸ”’

Functions

gather_named_type_vars ğŸ”’

type_mentions_data_type ğŸ”’